+++ paragraphs/para.json
@@ -1,0 +1,34 @@
+[
+  {
+    "id": "1",
+    "para" : "Hey man"
+  },
+  {
+    "id": "2",
+    "para" : "Hey woman"
+  },
+  {
+    "id": "3",
+    "para" : "Hey fuck"
+  },
+  {
+    "id": "4",
+    "para" : "Hey shit"
+  },
+  {
+    "id": "5",
+    "para" : "Hey nigga"
+  },
+  {
+    "id": "6",
+    "para" : "Hey whoa"
+  },
+  {
+    "id": "7",
+    "para" : "Hey mayne"
+  },
+  {
+    "id": "8",
+    "para" : "Hey brah"
+  }
+]
